# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * College Mart API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  college: string;
  year: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  college: string;
  year: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Item {
  id: string;
  name: string;
  price: number;
  category: string;
  description: string;
  condition?: string;
  sellerName: string;
  sellerWhatsapp: string;
  sellerId?: string;
  createdAt: string;
}

export interface ItemInput {
  name: string;
  price: number;
  category: string;
  description: string;
  condition?: string;
  sellerWhatsapp: string;
}

export interface ItemStats {
  total: number;
  categories: number;
  avgPrice: number;
}

export interface Demand {
  id: string;
  itemName: string;
  description: string;
  /** @nullable */
  budget?: number | null;
  posterName: string;
  posterWhatsapp: string;
  posterId?: string;
  createdAt: string;
}

export interface DemandInput {
  itemName: string;
  description: string;
  /** @nullable */
  budget?: number | null;
  posterWhatsapp: string;
}