# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: College Mart API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: items
    description: Stationery items for sale
  - name: demands
    description: Student demands

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /items:
    get:
      operationId: listItems
      tags: [items]
      summary: List all available stationery items
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"

    post:
      operationId: createItem
      tags: [items]
      summary: Upload a stationery item for sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemInput"
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /items/{id}:
    delete:
      operationId: deleteItem
      tags: [items]
      summary: Delete a stationery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /items/stats:
    get:
      operationId: getItemStats
      tags: [items]
      summary: Get item statistics
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemStats"

  /demands:
    get:
      operationId: listDemands
      tags: [demands]
      summary: List all demands
      responses:
        "200":
          description: List of demands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Demand"

    post:
      operationId: createDemand
      tags: [demands]
      summary: Post a demand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemandInput"
      responses:
        "201":
          description: Demand created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Demand"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /demands/{id}:
    delete:
      operationId: deleteDemand
      tags: [demands]
      summary: Delete a demand
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    RegisterInput:
      type: object
      required: [name, email, password, college, year]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        college:
          type: string
        year:
          type: string

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    User:
      type: object
      required: [id, name, email, college, year]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        college:
          type: string
        year:
          type: string

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    Item:
      type: object
      required: [id, name, price, category, description, sellerName, sellerWhatsapp, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
        category:
          type: string
        description:
          type: string
        condition:
          type: string
        sellerName:
          type: string
        sellerWhatsapp:
          type: string
        sellerId:
          type: string
        createdAt:
          type: string

    ItemInput:
      type: object
      required: [name, price, category, description, sellerWhatsapp]
      properties:
        name:
          type: string
        price:
          type: number
        category:
          type: string
        description:
          type: string
        condition:
          type: string
        sellerWhatsapp:
          type: string

    ItemStats:
      type: object
      required: [total, categories, avgPrice]
      properties:
        total:
          type: number
        categories:
          type: number
        avgPrice:
          type: number

    Demand:
      type: object
      required: [id, itemName, description, posterName, posterWhatsapp, createdAt]
      properties:
        id:
          type: string
        itemName:
          type: string
        description:
          type: string
        budget:
          type: ["number", "null"]
        posterName:
          type: string
        posterWhatsapp:
          type: string
        posterId:
          type: string
        createdAt:
          type: string

    DemandInput:
      type: object
      required: [itemName, description, posterWhatsapp]
      properties:
        itemName:
          type: string
        description:
          type: string
        budget:
          type: ["number", "null"]
        posterWhatsapp:
          type: string