# lib/api-zod/src/generated/types/demand.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * College Mart API
 * OpenAPI spec version: 0.1.0
 */

export interface Demand {
  id: string;
  itemName: string;
  description: string;
  /** @nullable */
  budget?: number | null;
  posterName: string;
  posterWhatsapp: string;
  posterId?: string;
  createdAt: string;
}