# lib/api-zod/src/generated/types/demandInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * College Mart API
 * OpenAPI spec version: 0.1.0
 */

export interface DemandInput {
  itemName: string;
  description: string;
  /** @nullable */
  budget?: number | null;
  posterWhatsapp: string;
}