# lib/api-zod/src/generated/types/item.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * College Mart API
 * OpenAPI spec version: 0.1.0
 */

export interface Item {
  id: string;
  name: string;
  price: number;
  category: string;
  description: string;
  condition?: string;
  sellerName: string;
  sellerWhatsapp: string;
  sellerId?: string;
  createdAt: string;
}