# lib/api-zod/src/generated/types/itemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * College Mart API
 * OpenAPI spec version: 0.1.0
 */

export interface ItemInput {
  name: string;
  price: number;
  category: string;
  description: string;
  condition?: string;
  sellerWhatsapp: string;
}