# lib/api-zod/src/generated/types/itemStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * College Mart API
 * OpenAPI spec version: 0.1.0
 */

export interface ItemStats {
  total: number;
  categories: number;
  avgPrice: number;
}